game-flow: move vars to libtrx
