+++ src/libtrx/game/game_flow/vars.c
