+++ src/libtrx/include/libtrx/game/game_flow/vars.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "./types.h"
+
+extern GAME_FLOW g_GameFlow;
