+++ src/tr1/meson.build
@@ -117,7 +117,6 @@ sources = [
   'game/game_flow/common.c',
   'game/game_flow/sequencer.c',
   'game/game_flow/sequencer_misc.c',
-  'game/game_flow/vars.c',
   'game/gun/gun.c',
   'game/gun/gun_misc.c',
   'game/gun/gun_pistols.c',
