--- src/tr2/game/game_flow/vars.c
@@ -1,4 +1,0 @@
-#include "game/game_flow/vars.h"
-
-GAME_FLOW g_GameFlow;
-GAME_INFO g_GameInfo;
