+++ src/tr2/game/game_flow/vars.h
@@ -1,6 +1,3 @@
 #pragma once
 
-#include "./types.h"
-
-extern GAME_FLOW g_GameFlow;
-extern GAME_INFO g_GameInfo;
+#include <libtrx/game/game_flow/vars.h>
