+++ src/tr2/meson.build
@@ -108,7 +108,6 @@ sources = [
   'game/game_flow/inventory.c',
   'game/game_flow/sequencer.c',
   'game/game_flow/sequencer_misc.c',
-  'game/game_flow/vars.c',
   'game/gun/gun.c',
   'game/gun/gun_misc.c',
   'game/gun/gun_pistols.c',
